% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\docType{package}
\name{biqr-package}
\alias{biqr}
\alias{biqr-package}
\title{Plot NOMe-seq data}
\description{
create bubble plot of NOMe-seq data
}
\details{
This package implements visualization of NOMeSeq data


}
\examples{
\dontrun{
library(biqr)
setwd('~/tools/biqr')
seq.fns <- paste0('1134071/', list.files('1134071', pattern=".seq"))
seq.fns.3F1R <- seq.fns[grepl("3F1R", seq.fns)]
seq.fns.2F2R <- seq.fns[grepl("2F2R", seq.fns)]
seq.fns.2F2R
seq.fns.3F1R

aln <- localaln(seq.fns.2F2R, "chr8", 11664775, 11667135)
aln <- localaln(seq.fns.3F1R, "chr8", 11664775, 11667135)
plotcytosine(aln$basemat[1:20,], 'hcg')
plotcytosine(aln$basemat[1:20,], 'gch')
}
}
\author{
Wanding Zhou \email{Wanding.Zhou@vai.org}
}

