% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{clone.aln}
\alias{clone.aln}
\title{Local align multiple clones}
\usage{
clone.aln(seq.fns, chrm, beg, end, min.score = 100, snames = NULL)
}
\arguments{
\item{seq.fns}{a vector of sequence file names}

\item{chrm}{target chromosome}

\item{beg}{target begin}

\item{end}{target end}

\item{min.score}{minimum alignment score for a query to be considered}

\item{snames}{a vector of sample names}
}
\value{
CloneAlignment
}
\description{
Local align multiple clones
}

