% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{clone.aln1}
\alias{clone.aln1}
\title{Local align one clone}
\usage{
clone.aln1(seq.fn, refseq)
}
\arguments{
\item{seq.fn}{sequence file name}

\item{refseq}{reference}
}
\value{
CloneAlignment1
}
\description{
Local align one clone
}

