% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{formatmat}
\alias{formatmat}
\title{Format a base matrix alignment}
\usage{
formatmat(m, col.width = 80)
}
\arguments{
\item{m}{base matrix for alignment}

\item{col.width}{column width}
}
\value{
a string
}
\description{
Format a base matrix alignment
}

