% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{get.reference}
\alias{get.reference}
\title{Get reference sequence}
\usage{
get.reference(chrm, beg, end)
}
\arguments{
\item{chrm}{chromosome}

\item{beg}{region start}

\item{end}{region end}
}
\value{
DNAString
}
\description{
Get reference sequence from DAS server
}

