% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{localaln}
\alias{localaln}
\title{Local align multiple queries}
\usage{
localaln(seq.fns, chrm, beg, end, min.score = 100)
}
\arguments{
\item{seq.fns}{a vector of sequence file names}

\item{chrm}{target chromosome}

\item{beg}{target begin}

\item{end}{target end}

\item{min.score}{minimum alignment score for a query to be considered}
}
\description{
Local align multiple queries
}

