% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{localaln1}
\alias{localaln1}
\title{Local align one query}
\usage{
localaln1(seq.fn, refseq)
}
\arguments{
\item{seq.fn}{sequence file name}

\item{refseq}{reference}
}
\value{
a list of aligned columns, score and strand
}
\description{
Local align one query
}

