% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{order.cytosine}
\alias{order.cytosine}
\title{Order clones by cytosine retention}
\usage{
order.cytosine(ca, context = "hcg", subset.c = TRUE)
}
\arguments{
\item{ca}{CloneAlignment}

\item{context}{"hcg"/"gch"/"cg"}

\item{subset.c}{whether to subset to cytosine of the specified context, default is TRUE}
}
\value{
a CloneAlignment (subset.c=FALSE) or a sorted base matrix (subset.c==TRUE)
}
\description{
Order clones by cytosine retention
}
\details{
This function only operates on the base matrix, but it requires the reference be intact (before subsetting).
}

