% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{plotcytosine}
\alias{plotcytosine}
\title{Plot cytosine}
\usage{
plotcytosine(basemat, target, margin = c(0.05, 0.05, 0.1, 0.05),
  bar.top = 0.05, alpha = 0.9)
}
\arguments{
\item{basemat}{base matrix of the alignment}

\item{target}{either "hcg", "gch" or "cg"}

\item{margin}{a vector of margins c(bottom, left, top, right)}

\item{bar.top}{height of top bar}

\item{alpha}{transparency}
}
\description{
Plot cytosine
}

