% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqr.R
\name{subsetCytosine}
\alias{subsetCytosine}
\title{Subset cytosine from CloneAlignment}
\usage{
subsetCytosine(m, context = "hcg")
}
\arguments{
\item{m}{base matrix}

\item{context}{"hcg" / "gch" / "cg"}
}
\value{
a matrix of cytosine alignment
}
\description{
Subset cytosine from CloneAlignment
}

